# api/openapi-spec/swagger.json
{
  "swagger": "2.0",
  "info": {
   "description": "Containerized Data Importer for KubeVirt.",
   "title": "KubeVirt Containerized Data Importer API",
   "contact": {
    "name": "kubevirt-dev",
    "url": "https://github.com/kubevirt/containerized-data-importer",
    "email": "kubevirt-dev@googlegroups.com"
   },
   "license": {
    "name": "Apache 2.0",
    "url": "https://www.apache.org/licenses/LICENSE-2.0"
   }
  },
  "paths": {
   "/": {
    "get": {
     "description": "Get a CDI API root paths",
     "produces": [
      "application/json"
     ],
     "operationId": "getUploadRootPaths",
     "responses": {
      "200": {
       "description": "OK",
       "schema": {
        "$ref": "#/definitions/v1.RootPaths"
       }
      },
      "401": {
       "description": "Unauthorized"
      },
      "404": {
       "description": "Not Found",
       "schema": {
        "type": "string"
       }
      }
     }
    }
   },
...
   "/apis/upload.cdi.kubevirt.io": {
    "get": {
     "description": "Get a CDI API Group",
     "produces": [
      "application/json"
     ],
     "operationId": "getUploadAPIGroup",
     "responses": {
      "200": {
       "description": "OK",
       "schema": {
        "$ref": "#/definitions/v1.APIGroup"
       }
      },
      "401": {
       "description": "Unauthorized"
      },
      "404": {
       "description": "Not Found",
       "schema": {
        "type": "string"
       }
      }
     }
    }
   },
...
  },
  "definitions": {
...
   "v1.APIGroup": {
    "description": "APIGroup contains the name, the supported versions, and the preferred version of a group.",
    "type": "object",
    "required": [
     "name",
     "versions"
    ],
    "properties": {
     "apiVersion": {
      "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
      "type": "string"
     },
     "kind": {
      "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
      "type": "string"
     },
     "name": {
      "description": "name is the name of the group.",
      "type": "string",
      "default": ""
     },
     "preferredVersion": {
      "description": "preferredVersion is the version preferred by the API server, which probably is the storage version.",
      "default": {},
      "$ref": "#/definitions/v1.GroupVersionForDiscovery"
     },
     "serverAddressByClientCIDRs": {
      "description": "a map of client CIDR to server address that is serving this group. This is to help clients reach servers in the most network-efficient way possible. Clients can use the appropriate server address as per the CIDR that they match. In case of multiple matches, clients should use the longest matching CIDR. The server returns only those CIDRs that it thinks that the client can match. For example: the master will return an internal IP CIDR only, if the client reaches the server using an internal IP. Server looks at X-Forwarded-For header or X-Real-Ip header or request.RemoteAddr (in that order) to get the client IP.",
      "type": "array",
      "items": {
       "default": {},
       "$ref": "#/definitions/v1.ServerAddressByClientCIDR"
      }
     },
     "versions": {
      "description": "versions are the versions supported in this group.",
      "type": "array",
      "items": {
       "default": {},
       "$ref": "#/definitions/v1.GroupVersionForDiscovery"
      }
     }
    }
   },
...
   "v1.GroupVersionForDiscovery": {
    "description": "GroupVersion contains the \"group/version\" and \"version\" string of a version. It is made a struct to keep extensibility.",
    "type": "object",
    "required": [
     "groupVersion",
     "version"
    ],
    "properties": {
     "groupVersion": {
      "description": "groupVersion specifies the API group and version in the form \"group/version\"",
      "type": "string",
      "default": ""
     },
     "version": {
      "description": "version specifies the version in the form of \"version\". This is to save the clients the trouble of splitting the GroupVersion.",
      "type": "string",
      "default": ""
     }
    }
   },
...
   "v1.RootPaths": {
    "description": "RootPaths lists the paths available at root. For example: \"/healthz\", \"/apis\".",
    "type": "object",
    "required": [
     "paths"
    ],
    "properties": {
     "paths": {
      "description": "paths are the paths available at root.",
      "type": "array",
      "items": {
       "type": "string",
       "default": ""
      }
     }
    }
   },
   "v1.ServerAddressByClientCIDR": {
    "description": "ServerAddressByClientCIDR helps the client to determine the server address that they should use, depending on the clientCIDR that they match.",
    "type": "object",
    "required": [
     "clientCIDR",
     "serverAddress"
    ],
    "properties": {
     "clientCIDR": {
      "description": "The CIDR with which clients can match their IP to figure out the server address that they should use.",
      "type": "string",
      "default": ""
     },
     "serverAddress": {
      "description": "Address of this server, suitable for a client that matches the above CIDR. This can be a hostname, hostname:port, IP or IP:port.",
      "type": "string",
      "default": ""
     }
    }
   },
...
  },
  "securityDefinitions": {
   "BearerToken": {
    "description": "Bearer Token authentication",
    "type": "apiKey",
    "name": "authorization",
    "in": "header"
   }
  }
 }